pyproject.toml
[tool.poetry]
name = "robot-test-creator"
version = "0.1.1"
description = ""
authors = ["Mika Hänninen <mika.hanninen@gmail.com>"]
license = "MIT"
repository = "https://github.com/mikahanninen/robot-test-creator"
keywords = ["robotframework", "tool"]

[tool.poetry.scripts]
rtc = "robot_test_creator.excel2suites:main"

[tool.poetry.dependencies]
python = "^3.7"
rpa-framework = "^0.10.1"
click = "^7.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
